Add support for devcontainers
Devcontainers make it very easy for project to distribute the working
development environment to their developers using VSCode.

This adds support for devcontainer, including the Github action.

The workflow is as following:

- With every new version of the container, go to the GitHub actions,
pick the "Build and publish devcontainer", and trigger the workflow
action with the new version as parameter

- Add the version to the devcontainer template, so that users can pick
it, also update the version of template so it gets published (just
increment the patch version), then trigger the workflow "Release
devcontainer template"